+++ CHANGELOG.md
@@ -248,6 +248,7 @@
 - Refactor middleware to use profile init @robgietema
 - Refactor seeds to use profile init @robgietema
 - Refactor behaviors to use profile init @robgietema
+- Refactor content rules to use profile init @robgietema
 
 ### Documentation